Fix conversion warning tripping

+++ source/states/main_menu_state.cpp
@@ -71,8 +71,8 @@ void MainMenuState::draw_state(float dt __attribute__((unused)))
     // set the text style
     text.setStyle(sf::Text::Bold | sf::Text::Underlined);
 
-    text.setPosition({(this->_data->window->getSize().x / 2) - (text.getLocalBounds().size.x / 2),
-                      (this->_data->window->getSize().y / 2) - (text.getLocalBounds().size.y / 2)});
+    text.setPosition({(static_cast<float>(this->_data->window->getSize().x / 2)) - (text.getLocalBounds().size.x / 2),
+                      (static_cast<float>(this->_data->window->getSize().y / 2)) - (text.getLocalBounds().size.y / 2)});
 
     // inside the main loop, between window.clear() and window.display()
     this->_data->window->draw(text);
